rm repeat code(can't in parse.go b/c of import cycle)
